Hong done search and role admin

+++ src/java/model/Restaurant.java
@@ -10,6 +10,7 @@ package model;
  * @author DELL
  */
 public class Restaurant {
+
     private int restaurantID;
     private String restaurantName;
     private int sellerID;
@@ -25,13 +26,14 @@ public class Restaurant {
     private float distance;
     private String description;
     private String restaurantImage;
-     private double starAVG;
+
+     private float starAVG;
 
     public double getStarAVG() {
         return starAVG;
     }
 
-    public void setStarAVG(double starAVG) {
+    public void setStarAVG(float starAVG) {
         this.starAVG = starAVG;
     }
     
@@ -124,7 +126,7 @@ public class Restaurant {
     }
 
     public void setCost(String cost) {
-        this.cost = cost ;
+        this.cost = cost;
     }
 
     public float getDistance() {
@@ -163,10 +165,8 @@ public class Restaurant {
     public void setRestaurantImage(String restaurantImage) {
         this.restaurantImage = restaurantImage;
     }
-    
-    
 
-    public Restaurant(String restaurantName, int sellerID, int countryID,int provinceID, int districtID, String address, String cost, float distance, String description) {
+    public Restaurant(String restaurantName, int sellerID, int countryID, int provinceID, int districtID, String address, String cost, float distance, String description) {
         this.restaurantName = restaurantName;
         this.sellerID = sellerID;
         this.countryID = countryID;
@@ -178,7 +178,7 @@ public class Restaurant {
         this.description = description;
     }
 
-    public Restaurant(int restaurantID, String restaurantName, String sellerName ,String countryName, String provinceName, String districtName, String address, String cost, float distance, String description,String restaurantImage ) {
+    public Restaurant(int restaurantID, String restaurantName, String sellerName, String countryName, String provinceName, String districtName, String address, String cost, float distance, String description, String restaurantImage) {
         this.restaurantID = restaurantID;
         this.restaurantName = restaurantName;
         this.sellerName = sellerName;
@@ -192,7 +192,7 @@ public class Restaurant {
         this.restaurantImage=restaurantImage;
         
     }
-    public Restaurant(int restaurantID, String restaurantName, String sellerName ,String countryName, String provinceName, String districtName, String address, String cost, float distance, String description,String restaurantImage, double starAVG ) {
+    public Restaurant(int restaurantID, String restaurantName, String sellerName ,String countryName, String provinceName, String districtName, String address, String cost, float distance, String description,String restaurantImage, float starAVG ) {
         this.restaurantID = restaurantID;
         this.restaurantName = restaurantName;
         this.sellerName = sellerName;
@@ -207,7 +207,7 @@ public class Restaurant {
         this.starAVG = starAVG;
     }
 
-    public Restaurant(int restaurantID, String restaurantName, int sellerID, int countryID, int provinceID, int districtID, String address, String cost, float distance, String description,String restaurantImage) {
+    public Restaurant(int restaurantID, String restaurantName, int sellerID, int countryID, int provinceID, int districtID, String address, String cost, float distance, String description, String restaurantImage) {
         this.restaurantID = restaurantID;
         this.restaurantName = restaurantName;
         this.sellerID = sellerID;
@@ -218,10 +218,7 @@ public class Restaurant {
         this.cost = cost;
         this.distance = distance;
         this.description = description;
-        this.restaurantImage=restaurantImage;
+        this.restaurantImage = restaurantImage;
     }
 
-   
-    
-    
 }
